+++ pokedex/src/components/home/Home.js
@@ -18,48 +18,20 @@ const Pokecard = styled.div`
   gap: 10px;
 `;
 
-const Home = () => {
-  const [pokemonlist, setPokemonlist] = useState([]);
-
-  useEffect(() => {
-    axios
-      .get("https://pokeapi.co/api/v2/pokemon")
-      .then((res) => {
-        setPokemonlist(res.data.results);
-        console.log(res.data.results)
-        console.log(pokemon)
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }, []);
-
-  const buildImgUrl = (url) => {
-    const id = url.split("/");
-    const idx = id.length - 2;
-    const imgUrl = `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${id[idx]}.png`;
-
-    return imgUrl;
-    };
+const Home = (props) => {
   
-    const getPokemonId = (url) => {
-      const id = url.split("/");
-      const idx = id.length - 2;
-      const pokemonId = id[idx];
-      return pokemonId;
-    };
 
 
   return (
     
     <PokecardContainer>
-      {pokemonlist.map((pokemon) => ( 
+      {props.pokemonlist.map((pokemon) => ( 
             
         <Pokecard key={pokemon.name}>
             <h2>{pokemon.name}</h2>
-            <img src={buildImgUrl(pokemon.url)} alt={pokemon.name} />
+            <img src={props.buildImgUrl(pokemon.url)} alt={pokemon.name} />
               <button>Adicionar a pokedex</button>
-              <Link to={`/details/${getPokemonId(pokemon.url)}`}>  
+              <Link to={`/details/${props.getPokemonId(pokemon.url)}`}>  
               <button>Ver detalhes</button>  </Link>
                         
         </Pokecard>
